refactored ConvertSemanticToken
